requirements.txt
ecdsa==0.19.0
keyboard==0.13.5
pycryptodome==3.20.0
six==1.16.0
